refactor: cahnged folder name for proper command
changed bundles to bundle in order to generate proper command
