+++ schemas/package-bundle-create.json
@@ -1,0 +1,16 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$ref": "#/definitions/BundleCreate",
+  "definitions": {
+    "BundleCreate": {
+      "type": "object",
+      "properties": {
+        "Id": {
+          "type": "string"
+        }
+      },
+      "required": ["Id"],
+      "additionalProperties": false
+    }
+  }
+}
