+++ schemas/package-bundle-list.json
@@ -1,0 +1,58 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "$ref": "#/definitions/BundleListCommandResults",
+  "definitions": {
+    "BundleListCommandResults": {
+      "type": "array",
+      "items": {
+        "$ref": "#/definitions/BundleListCommandResult"
+      }
+    },
+    "BundleListCommandResult": {
+      "$ref": "#/definitions/BundleSObjects.Bundle"
+    },
+    "BundleSObjects.Bundle": {
+      "type": "object",
+      "properties": {
+        "BundleName": {
+          "type": "string"
+        },
+        "Description": {
+          "type": "string"
+        },
+        "Id": {
+          "type": "string"
+        },
+        "IsDeleted": {
+          "type": "boolean"
+        },
+        "CreatedDate": {
+          "type": "string"
+        },
+        "CreatedById": {
+          "type": "string"
+        },
+        "LastModifiedDate": {
+          "type": "string"
+        },
+        "LastModifiedById": {
+          "type": "string"
+        },
+        "SystemModstamp": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "BundleName",
+        "Id",
+        "IsDeleted",
+        "CreatedDate",
+        "CreatedById",
+        "LastModifiedDate",
+        "LastModifiedById",
+        "SystemModstamp"
+      ],
+      "additionalProperties": false
+    }
+  }
+}
