+++ src/commands/package/bundle/list.ts
@@ -24,7 +24,7 @@ export class BundleListCommand extends SfCommand<BundleListCommandResults> {
   public static readonly description = messages.getMessage('description');
   public static readonly examples = messages.getMessages('examples');
   public static readonly deprecateAliases = true;
-  public static readonly aliases = ['force:bundle:list'];
+  public static readonly aliases = ['force:package:bundle:list'];
   public static readonly flags = {
     loglevel,
     'target-dev-hub': requiredHubFlag,
